refactor(about-us): extract `screen` out as a constant
