+++ src/components/about-us/section-02/content-carousel-list.js
@@ -2,16 +2,17 @@ import Arrows from './arrows'
 import Navigation from '../utils/navigation'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
+import carouselMarkup from '../constants/section-02/carousel-markup' 
 import categories from '../constants/section-02/categories'
 import categoryIds from '../constants/section-02/category-ids'
-import carouselMarkup from '../constants/section-02/carousel-markup' 
 import colors from '../../../constants/colors'
 import mq from '../utils/media-query'
+import screen from '../utils/screen'
 import styled from 'styled-components'
 import { font } from '../constants/styles'
 import { gray } from './utils'
+import { headcountPerPage } from '../constants/section-02/headcount-per-page'
 import { replaceGCSUrlOrigin } from '@twreporter/core/lib/utils/storage-url-processor'
-import { screen, headcountPerPage } from '../constants/section-02/headcount-per-page'
 import { storageUrlPrefix } from '../utils/config'
 //lodash
 import assign from 'lodash/assign'
