+++ src/components/about-us/section-02/content-paginated-list.js
@@ -7,10 +7,11 @@ import categories from '../constants/section-02/categories'
 import categoryIds from '../constants/section-02/category-ids'
 import colors from '../../../constants/colors'
 import mq from '../utils/media-query'
+import screen from '../utils/screen'
 import styled from 'styled-components'
 import values from 'lodash/values'
 import { gray } from './utils'
-import { headcountPerPage, screen } from '../constants/section-02/headcount-per-page'
+import { headcountPerPage } from '../constants/section-02/headcount-per-page'
 import { replaceGCSUrlOrigin } from '@twreporter/core/lib/utils/storage-url-processor'
 import { storageUrlPrefix } from '../utils/config'
