+++ src/components/about-us/utils/screen.js
@@ -1,0 +1,9 @@
+const screen = {
+  hd: 'hd',
+  desktop: 'desktop',
+  tablet: 'tablet',
+  mobile: 'mobile'
+}
+
+export default screen
+
